putting some validations on folder path

+++ app/js/script.js
@@ -4,29 +4,36 @@ $(function(){
     var watcher = null;
     
     watchFolder = function(folderpath){
-        watcher = fs.watch(folderpath, (event, filename) => {
-          if (filename) {
-            fs.exists(folderpath+'\\'+filename, (exists) => {
-                if (exists && event === 'rename'){
-                    fs.unlinkSync(folderpath+filename);
-                    $('#textlog').append('<li>processando: '+folderpath+'\\'+filename+'</li>');
-                }
+        
+        if (fs.exists(folderpath) && fs.lstatSync(folderpath).isDirectory()){
+        
+            watcher = fs.watch(folderpath, (event, filename) => {
+              if (filename) {
+                fs.exists(folderpath+'\\'+filename, (exists) => {
+                    if (exists && event === 'rename'){
+                        fs.unlinkSync(folderpath+filename);
+                        $('#textlog').append('<li>processando: '+folderpath+'\\'+filename+'</li>');
+                    }
+                });
+              }
             });
             
-          } else {
-            console.log('filename not provided');
-          }
-        });
+            return true;
+        }else{
+            alert('invalid folder path '+folderpath);
+            return false;
+        }
     };
 
     $('#watch').click(function(e){
         var folder = $('#folder').val();
-        
+
         if (watcher){
             watcher.close();
         }
         
-        watchFolder('C:\\temp\\deleteme\\');
-        $('#textlog').append('<li>monitorando: '+folder+'</li>');
+        if (watchFolder(folder)){
+            $('#textlog').append('<li>monitorando: '+folder+'</li>');
+        }
     });
 });
